zerr in one object

+++ src/modules/featurebank.cpp
@@ -33,7 +33,7 @@ void FeatureBank::print_all_features(){
 
 void FeatureBank::print_active_features(){
     std::cout<<"All activated features: "<<std::endl;
-    for (int i = 0; i < activated_features.size(); ++i){
+    for (size_t i = 0; i < activated_features.size(); ++i){
         std::cout<<"  -Name: "<<activated_features[i]->get_name()<<std::endl;
         std::cout<<"  -Category: "<<activated_features[i]->get_category()<<std::endl;
         std::cout<<"  -Description: "<<activated_features[i]->get_description()<<std::endl;
